pyproject.toml
[tool.poetry]
name = "tweet-monitor"
version = "0.1.0"
description = ""
authors = ["shin-hama <htysivista@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tweepy = "^4.12.1"
python-dotenv = "^0.21.0"
line-bot-sdk = "^2.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.991"
flake8 = "^6.0.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
